requirements.txt
certifi==2020.12.5
elasticsearch==7.11.0
elasticsearch-dsl==7.3.0
humanize==3.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
urllib3==1.26.3
